+++ Unity Project Test/Assets/Scripts/BallBehaviour/PlayerCamera.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Camera : MonoBehaviour {
+public class PlayerCamera : MonoBehaviour {
 
     public GameObject player;
