+++ Unity Project Test/Assets/Scripts/CubeBehaviour.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5fde6064e707f9d4fa390c03d5a8e497
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
