+++ Unity Project Test/Assets/Scripts/CubeBehaviour/LocalCube.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 735a6dcf2bb5e82498fd52f2502d1f9a
+guid: 4be9f2d66ea1eee4b9ec5c921676542f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
