+++ Unity Project Test/Assets/Scripts/CubeBehaviour/NetworkCube.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c9a30a4cadaa41208d46fb226ea5853c
+guid: d00e7937d27a50242819ce0744ea0ec2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
